backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-storages[boto3]>=1.14
celery[redis]>=5.3
psycopg2-binary>=2.9
PyPDF2>=3.0
scikit-learn>=1.4
nltk>=3.8
gunicorn>=21.2
python-dotenv>=1.0
easyocr>=1.7
opencv-python-headless>=4.8
